refactor: create get and move element modules
