+++ packages/core/index.js
@@ -2,9 +2,10 @@ export * from './src/js/accessibility';
 export * from './src/js/addClass';
 export * from './src/js/camelCase';
 export * from './src/js/focus';
-export * from './src/js/focusTrap';
+export * from './src/js/getElement';
 export * from './src/js/hasClass';
 export * from './src/js/hyphenCase';
+export * from './src/js/moveElement';
 export * from './src/js/removeClass';
 export * from './src/js/toggleClass';
 export * from './src/js/transition';
