+++ packages/core/src/js/variables.js
@@ -1,11 +1,7 @@
-const breakpoints = {
+export const breakpoints = {
   xs: '480px',
   sm: '620px',
   md: '760px',
   lg: '990px',
   xl: '1380px'
-};
-
-export {
-  breakpoints
 };
